host-company-research/requirements.txt
# Company Research Agent
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
tavily-python>=0.5.0
pydantic>=2.0.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0

